requirements.txt
pyOpenSSL==0.13